Muchos cambios (autenticación de usuario nos permitió avanzar con muchas funcionalidades)

+++ src/services/servicio_reservas.ts
@@ -1,7 +1,7 @@
 import { Reserva } from '@prisma/client';
 import { prisma } from '../index';
 
-export class ServicioPlato {
+export class ServicioReserva {
     // Admin
     static async obtener_reservas(): Promise<Reserva[]> {
         return prisma.reserva.findMany();
@@ -16,37 +16,46 @@ export class ServicioPlato {
         });
     }
 
-    // Admin?
-    static async obtener_reservas_fecha(fecha: Date): Promise<Reserva[]> {
-        return prisma.reserva.findMany({
-            where: {
-                fecha: fecha,
-            },
-        });
-    }
-
     // Admin
     static async crear_reserva(datos: Omit<Reserva, 'id' | 'usuario' | 'mesa'>): Promise<Reserva> {
         const reserva = await prisma.reserva.create({
-            data: {
-                usuarioId: datos.usuarioId,
-                mesaId: datos.mesaId,
-                fecha: datos.fecha,
-            },
+            data: datos
         });
-    
+
         await prisma.mesa.update({
-            where: { id: datos.mesaId },
-            data: { disponible: false },
-        });
+            where: {
+                id: reserva.mesaId
+            },
+            data: {
+                disponible: false
+            }
+        })
     
         return reserva;
     }
 
     // Admin
-    static async eliminar_reserva(id: number): Promise<Reserva> {
-        return prisma.reserva.delete({
+    static async eliminar_reserva(id: number) {
+        const reserva = await prisma.reserva.findUnique({
             where: { id },
         });
+
+        if (!reserva){
+            throw new Error('Reserva no encontrada');
+        }
+
+        await prisma.reserva.delete({
+            where: { id },
+        });
+
+        await prisma.mesa.update({
+            where: {
+                id: (await reserva).mesaId
+            },
+            data: {
+                disponible: true
+            }
+        })
+
     }
 }
